Frontend: Big fat remake
