--- frontend/lobby/join/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width,initial-scale=1.0">
-        <title>Troll Card Game</title>
-    </head>
-    <body>
-        <script src="/config/js-yaml.min.js"></script>
-        <script src="/config/config.js"></script>
-        <script src="join.js"></script>
-    </body>
-</html>
