--- frontend/lobby/waiting/index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width,initial-scale=1.0">
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
-        <title>Troll Card Game</title>
-    </head>
-    <body>
-        <div id="app" class="text-center container">
-            <h1 class="display-1">
-                <a href="/">
-                    TCG
-                </a>
-            </h1>
-            <div id="waiting-for-opponent" class="row">
-                <div id="game-status" class="col-12">
-                </div>
-            </div>
-            <template id="joinTemplate">
-                <div id="join-link" class="col-12">
-                </div>
-            </template>
-        </div>
-        <script src="/config/js-yaml.min.js"></script>
-        <script src="/config/config.js"></script>
-        <script src="waiting.js"></script>
-    </body>
-</html>
