+++ frontend/package.json
@@ -1,0 +1,21 @@
+{
+  "license": "MIT",
+  "dependencies": {
+    "clean-webpack-plugin": "^0.1.19",
+    "html-webpack-plugin": "^3.0.6",
+    "json-loader": "^0.5.7",
+    "node-fetch": "^2.1.1",
+    "pug": "^2.0.1",
+    "pug-loader": "git://github.com/pugjs/pug-loader.git",
+    "webpack": "^4.1.1",
+    "yaml-loader": "^0.5.0"
+  },
+  "scripts": {
+    "build": "webpack",
+    "start": "webpack-dev-server"
+  },
+  "devDependencies": {
+    "webpack-cli": "^2.0.12",
+    "webpack-dev-server": "^3.1.1"
+  }
+}
