+++ frontend/src/config/config.js
@@ -1,0 +1,1 @@
+module.exports = require('./data/config.yml')
