+++ frontend/src/game/cleanup.js
@@ -1,0 +1,9 @@
+"use strict";
+
+module.exports = {
+    clearChildren(elem) {
+        while (elem.firstChild) {
+            elem.removeChild(elem.firstChild);
+        }
+    },
+}
