+++ frontend/src/game/game.pug
@@ -1,0 +1,33 @@
+extends ../tcg
+
+block view
+    #status-row.row
+        #game-status.col-12
+    #game
+        #board.row
+            #my-score.col.text-right 0
+            #opponent-score.col.text-left 0
+            .w-100
+            #rows.col-12
+                #row-card-title.row
+                    #my-game-row-title.col-6.text-center
+                        h2 My cards
+                    #opponent-game-row-title.col-6.text-center
+                        h2 Opponent's cards
+                .row
+                    #my-rows.col-6
+                    #opp-rows.col-6
+            .col-12
+                .hand.row
+    template#row-template
+        .game-row.row.text-center.justify-content-end.border(
+            ondragover!="allowDrop(event)"
+            ondrop!="dropOnRow(event)"
+            onclick!="clickOnRow(event)"
+            style="height: 50px"
+            )
+    template#card-template
+        .card.col-1(
+            draggable="true"
+            ondragstart!="dragCardFromHand(event)"
+        )
