+++ frontend/src/game/templates.js
@@ -1,0 +1,6 @@
+"use strict";
+
+module.exports = {
+    baseCard: document.getElementById("card-template")
+            .content.querySelector(".card")
+}
