+++ frontend/src/index.pug
@@ -1,0 +1,6 @@
+extends tcg
+
+block view
+    .main-view
+        a#create-game.btn.btn-primary(href="/lobby_create.html")
+            | Create Game
