+++ frontend/src/lobby/create/create.pug
@@ -1,0 +1,4 @@
+extends ../../base
+
+block title
+    title TCG - Creating Game
