+++ frontend/src/lobby/join/join.pug
@@ -1,0 +1,4 @@
+extends ../../base
+
+block title
+    title TCG - Join Game
