+++ frontend/src/lobby/waiting/waiting.pug
@@ -1,0 +1,7 @@
+extends ../../tcg
+
+block view
+    #waiting-for-opponent.row
+        #game-status.col-12
+    template#joinTemplate
+        #join-link.col-12
