+++ frontend/src/tcg.pug
@@ -1,0 +1,10 @@
+extends base
+
+block title
+    title TCG
+
+block content
+    #app.text-center.container
+        h1.display-1
+            a(href="/") TCG
+        block view
